Comentarios

+++ src/business/SongManager.java
@@ -13,28 +13,47 @@ import java.io.IOException;
 import java.util.LinkedList;
 import java.util.Scanner;
 
-public class SongManager<Public> {
+/**
+ * Clase que gestiona la interaccion de las canciones con la base de datos
+ */
+public class SongManager {
 
     private static SongDAO songDAO = new SQLConnectorSong();
 
     private static SongPlaylistDAO songPDAO = new SQLConnectorSongPlaylist();
 
     /**
-     * Metodo que se utiliza para añadir canciones y se comunica con los sqlconnectors para añadir canciones a la base de datos.
-     * @param title
-     * @param Genre
-     * @param album
-     * @param artist
-     * @param path
-     * @param owner
+     * Método que añade una canción
+     * @param title Título de una canción
+     * @param Genre Género de una canción
+     * @param album Album al que pertenece la canción
+     * @param artist Artista de la canción
+     * @param path Path donde se encuentra la canción
+     * @param owner Usuario que añadió la canción
      */
     public static void addSong (String title, String Genre, String album, String artist, String path,String owner) {
         songDAO.InsertDataSong(title, Genre, album, artist, path, owner);
     }
 
     /**
-     * Metodo que llama a los sqlconnectors para retornar una linked list con todas las canciones de la base de datos.
-     * @return
+     * Mira si la cancion existe en la base de datos
+     * @param title Título de la canción
+     * @param path Path donde se encuentra la canción
+     * @return true si la canción existe
+     */
+    public static boolean checkSongExistance (String title,String path) {
+        LinkedList<Song> songs = ListSongs();
+        for (Song s: songs) {
+            if (s.getTitle().equals(title) || s.getPath().equals(path)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * Método que devuelve una LinkedList con las canciones cargadas
+     * @return LinkedList canciones
      */
     public static LinkedList<Song> ListSongs() {
         songDAO = new SQLConnectorSong();
@@ -45,29 +64,14 @@ public class SongManager<Public> {
     }
 
     /**
-     * Metodo que inserta nuevas canciones en la base de datos llamando a los metodos pertinentes del sqlconnector song
-     * @param title
-     * @param genre
-     * @param album
-     * @param artist
-     * @param path
-     * @param owner
+     * Método que retorna una canción segun el nombre de la canción
+     * @param name Nombre de la canción
+     * @return Song Cancion encontrada
      */
-    public static void InsertNewSong(String title, String genre, String album, String artist, String path, String owner){
-        songDAO = new SQLConnectorSong();
-
-        songDAO.InsertDataSong(title, genre, album, artist, path, owner);
-    }
-
-    /**
-     * Metodo que dado el nombre de una cancion devuelve toda su informacion
-     * @param Song
-     * @return
-     */
-    public static Song SelectSong(String Song) {
+    public static Song SelectSong(String name) {
         songDAO = new SQLConnectorSong();
         Song songs;
-        songs = songDAO.SelectSong(Song);
+        songs = songDAO.SelectSong(name);
 
         System.out.println(songs.getGenre());
 
@@ -75,26 +79,21 @@ public class SongManager<Public> {
     }
 
     /**
-     * Metodo que se utiliza para eliminar una cancion de la base de datos tanto en la tabla song como en la tabla song_playlist.
-
-     * @param title
+     * Método que elimina una canción
+     * @param title Titulo de una cancion
+     * @return true si la cancion ha sido eliminada
+     *         false si la cancion no se pudo eliminar
      */
-    public static void DeleteSong(String title){
-        songDAO.DeleteDataSong(title);
-        songPDAO.DeleteSongPFull(title);
-
+    public static boolean DeleteSong(String title){
+        for (Song song: ListSongs()) {
+            if (song.getTitle().equals(title)) {
+                songDAO.DeleteDataSong(title);
+                songPDAO.DeleteSongPFull(title);
+                return true;
+            }
+        }
+        return false;
     }
-    /*public static LinkedList<Song> SelectSong(String Song) {
-        songDAO = new SQLConnectorSong();
-        LinkedList<Song> song;
-        song = songDAO.SelectDataSong();
-        System.out.println(song);
-        return song;
-    }*/
-
-
-
-
 }
